+++ include/GafferSceneTest/TestShader.h
@@ -54,6 +54,11 @@ class GAFFERSCENETEST_API TestShader : public GafferScene::Shader
 
 		GAFFER_NODE_DECLARE_TYPE( GafferSceneTest::TestShader, TestShaderTypeId, GafferScene::Shader );
 
+		// Populates the `parameters` with plugs for different test cases. Currently
+		// supports `simpleShader` and `simpleLight`. If `shaderName` is not recognized,
+		// it will create a shader with that name and no parameters.
+		void loadShader( const std::string &shaderName, bool keepExistingValues=false ) override;
+
 };
 
 } // namespace GafferSceneTest