+++ python/GafferArnoldTest/RenderPassAdaptorTest.py
@@ -1,6 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2023, Cinesite VFX Ltd. All rights reserved.
+#  Copyright (c) 2024, Cinesite VFX Ltd. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -34,23 +34,32 @@
 #
 ##########################################################################
 
-import IECore
-import Gaffer
+import unittest
 
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "label", "Enabled" )
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "description", "Whether the render pass is enabled for rendering." )
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "defaultValue", IECore.BoolData( True ) )
+import GafferSceneTest
+import GafferArnold
 
-Gaffer.Metadata.registerValue( "option:renderPass:type", "label", "Type" )
-Gaffer.Metadata.registerValue(
-	"option:renderPass:type",
-	"description",
-	"""
-	The type of the render pass. This provides simple setup for renders such as reflection and shadow passes,
-	typically by assigning custom shaders to the objects specified by `Casters` and `Catchers`. Use a RenderPassShaders
-	node to customise the shaders used for this purpose.
+class RenderPassAdaptorTest( GafferSceneTest.RenderPassAdaptorTest ) :
 
-	> Hint : Render pass types and their behaviours can be customised using the RenderPassTypeAdaptor API.
-	"""
-)
-Gaffer.Metadata.registerValue( "option:renderPass:type", "defaultValue", IECore.StringData( "" ) )
+	renderer = "Arnold"
+
+	def _createDistantLight( self ) :
+
+		light = GafferArnold.ArnoldLight()
+		light.loadShader( "distant_light" )
+		return light, light["parameters"]["color"]
+
+	def _createStandardShader( self ) :
+
+		shader = GafferArnold.ArnoldShader()
+		shader.loadShader( "standard_surface" )
+		return shader, shader["parameters"]["base_color"]
+
+	def _createFlatShader( self ) :
+
+		shader = GafferArnold.ArnoldShader()
+		shader.loadShader( "flat" )
+		return shader, shader["parameters"]["color"]
+
+if __name__ == "__main__":
+	unittest.main()