+++ python/GafferSceneTest/RenderPassShaderTest.py
@@ -1,6 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2023, Cinesite VFX Ltd. All rights reserved.
+#  Copyright (c) 2024, Cinesite VFX Ltd. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -34,23 +34,26 @@
 #
 ##########################################################################
 
-import IECore
 import Gaffer
+import GafferScene
+import GafferSceneTest
 
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "label", "Enabled" )
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "description", "Whether the render pass is enabled for rendering." )
-Gaffer.Metadata.registerValue( "option:renderPass:enabled", "defaultValue", IECore.BoolData( True ) )
+class RenderPassShaderTest( GafferSceneTest.SceneTestCase ) :
 
-Gaffer.Metadata.registerValue( "option:renderPass:type", "label", "Type" )
-Gaffer.Metadata.registerValue(
-	"option:renderPass:type",
-	"description",
-	"""
-	The type of the render pass. This provides simple setup for renders such as reflection and shadow passes,
-	typically by assigning custom shaders to the objects specified by `Casters` and `Catchers`. Use a RenderPassShaders
-	node to customise the shaders used for this purpose.
+	def testAssignment( self ) :
 
-	> Hint : Render pass types and their behaviours can be customised using the RenderPassTypeAdaptor API.
-	"""
-)
-Gaffer.Metadata.registerValue( "option:renderPass:type", "defaultValue", IECore.StringData( "" ) )
+		s = GafferSceneTest.TestShader()
+		r = GafferScene.RenderPassShader()
+		r["shader"].setInput( s["out"] )
+
+		for usage, renderer in (
+			( "test", "Arnold" ),
+			( "shadow", "Cycles" ),
+			( "reflection", "*" )
+		) :
+			r["usage"].setValue( usage )
+			r["renderer"].setValue( renderer )
+			self.assertTrue( "i" in r["out"]["globals"].getValue()[f"option:renderPass:shader:{usage}:{renderer}"].outputShader().parameters )
+
+if __name__ == "__main__":
+	unittest.main()