+++ python/GafferUI/PlugLayout.py
@@ -357,7 +357,10 @@ class PlugLayout( GafferUI.Widget ) :
 			# Must scope the context because summaries are typically
 			# generated from plug values, and may therefore trigger
 			# a compute.
-			section.summary = self.__metadataValue( self.__parent, self.__layoutName + ":section:" + section.fullName + ":summary" ) or ""
+			try :
+				section.summary = self.__metadataValue( self.__parent, self.__layoutName + ":section:" + section.fullName + ":summary" ) or ""
+			except Gaffer.ProcessException :
+				section.summary = "<img src={}>".format( Gaffer.rootPath() / "graphics" / "errorSmall.png" )
 
 		section.valuesChanged = False