+++ python/IECoreDelightTest/RendererTest.py
@@ -46,6 +46,8 @@ import math
 
 import imath
 
+import OpenImageIO
+
 import IECore
 import IECoreScene
 import IECoreDelight
@@ -91,7 +93,7 @@ class RendererTest( GafferTest.TestCase ) :
 				"rgba",
 				{
 					"filter" : "gaussian",
-					"filterwidth" : imath.V2f( 3.5 ),
+					"filterwidth" : 3.5,
 				}
 			)
 		)
@@ -100,6 +102,42 @@ class RendererTest( GafferTest.TestCase ) :
 		del r
 
 		self.assertTrue( ( self.temporaryDirectory() / "beauty.exr" ).exists() )
+
+	def testOutputMultipart( self ) :
+
+		r = GafferScene.Private.IECoreScenePreview.Renderer.create(
+			"3Delight",
+			GafferScene.Private.IECoreScenePreview.Renderer.RenderType.Batch
+		)
+
+		r.output(
+			"layer_a",
+			IECoreScene.Output(
+				str( self.temporaryDirectory() / "multipart.exr" ),
+				"exr",
+				"rgba"
+			)
+		)
+
+		r.output(
+			"layer_b",
+			IECoreScene.Output(
+				str( self.temporaryDirectory() / "multipart.exr" ),
+				"exr",
+				"color shader:diffuse"
+			)
+		)
+
+		r.render()
+		del r
+
+		self.assertTrue( ( self.temporaryDirectory() / "multipart.exr" ).exists() )
+
+		i = OpenImageIO.ImageInput.open( os.path.join( self.temporaryDirectory(), "multipart.exr" ) )
+		subimages = i.spec().getattribute("oiio:subimages")
+		del i
+
+		self.assertEqual( subimages, 2 )
 
 	def testAOVs( self ) :
 
@@ -156,7 +194,7 @@ class RendererTest( GafferTest.TestCase ) :
 					data,
 					{
 						"filter" : "gaussian",
-						"filterwidth" : imath.V2f( 3.5 ),
+						"filterwidth" : 3.5,
 					}
 				)
 			)
@@ -1540,6 +1578,56 @@ class RendererTest( GafferTest.TestCase ) :
 		self.assertEqual( nsi["outputLayer:diffuse"]["layername"], "diffuse" )
 		self.assertEqual( nsi["outputLayer:directDiffuse"]["layername"], "diffuse_direct" )
 		self.assertEqual( nsi["outputLayer:customLayerName"]["layername"], "myLayerName" )
+
+	def testOutputLayerAttributes( self ) :
+
+		for data, expected in {
+			"rgba" : {
+				"variablename": "Ci",
+				"variablesource": "shader",
+				"layertype": "color",
+				"withalpha": 1,
+				"filter": "gaussian",
+				"filterwidth": 3.5,
+				"scalarformat": "half",
+				"colorprofile": "sRGB",
+				"layername": "Test",
+				"dithering": 1
+			},
+		}.items() :
+
+			r = GafferScene.Private.IECoreScenePreview.Renderer.create(
+				"3Delight",
+				GafferScene.Private.IECoreScenePreview.Renderer.RenderType.SceneDescription,
+				str( self.temporaryDirectory() / "test.nsia" )
+			)
+
+			r.output(
+				"test",
+				IECoreScene.Output(
+					"beauty.exr",
+					"exr",
+					data,
+					{
+						"filter": "gaussian",
+						"filterwidth": 3.5,
+						"scalarformat": "half",
+						"colorprofile": "sRGB",
+						"layerName": "Test",
+						"dithering": 1
+					}
+				)
+			)
+
+			r.render()
+			del r
+
+			nsi = self.__parseDict( self.temporaryDirectory() / "test.nsia" )
+			self.assertIn( "outputLayer:test", nsi )
+			self.assertEqual( nsi["outputLayer:test"]["nodeType"], "outputlayer")
+			for k, v in expected.items() :
+				self.assertIn( k, nsi["outputLayer:test"] )
+				self.assertEqual( nsi["outputLayer:test"][k], v )
 
 	# Helper methods used to check that NSI files we write contain what we
 	# expect. The 3delight API only allows values to be set, not queried,